fix: Complete audiobook compilation fixes
- Restore MainActivity from 7498628
- Add audiobook navigation routes and imports
- Replace AI service with text fallback
- Fix OfflineAudioPlayer seekTo redundancy
- Fix PlaybackControls callback syntax

+++ app/src/main/java/com/secondmind/minimal/utils/OfflineAudioPlayer.kt
@@ -47,7 +47,7 @@ class OfflineAudioPlayer {
     }
 
     fun seekTo(position: Int) {
-        player?.seekTo(position.toInt())
+        player?.seekTo(position)
     }
 
     fun setSpeed(speed: Float) {
